+++ apps/studio/components/interfaces/Database/Replication/ReplicationPipelineStatus/ReplicationPipelineStatus.utils.tsx
@@ -1,9 +1,14 @@
-import { ReplicationPipelineStatusData } from 'data/replication/pipeline-status-query'
+import dayjs from 'dayjs'
 import { Activity, Clock, HelpCircle, Loader2, XCircle } from 'lucide-react'
+
+import { ReplicationPipelineStatusData } from 'data/replication/pipeline-status-query'
+import { formatBytes } from 'lib/helpers'
 import { PipelineStatusRequestStatus } from 'state/replication-pipeline-request-status'
 import { Badge } from 'ui'
-import { getPipelineStateMessages } from './Pipeline.utils'
+import { getPipelineStateMessages } from '../Pipeline.utils'
 import { RetryPolicy, TableState } from './ReplicationPipelineStatus.types'
+
+const numberFormatter = new Intl.NumberFormat()
 
 export const getStatusConfig = (state: TableState['state']) => {
   switch (state.name) {
@@ -128,3 +133,57 @@ export const isValidRetryPolicy = (policy: any): policy is RetryPolicy => {
       return false
   }
 }
+
+const formatLagBytesValue = (value?: number) => {
+  if (typeof value !== 'number' || Number.isNaN(value)) {
+    return { display: '—', detail: undefined }
+  }
+
+  const decimals = value < 1024 ? 0 : value < 1024 * 1024 ? 1 : 2
+  const display = formatBytes(value, decimals)
+  const detail = `${numberFormatter.format(value)} bytes`
+
+  return { display, detail }
+}
+
+const formatLagDurationValue = (value?: number) => {
+  if (typeof value !== 'number' || Number.isNaN(value)) {
+    return { display: '—', detail: undefined }
+  }
+
+  const sign = value < 0 ? '-' : ''
+  const absMilliseconds = Math.abs(value)
+  const duration = dayjs.duration(absMilliseconds, 'milliseconds')
+
+  if (absMilliseconds < 1000) {
+    return { display: `${value} ms`, detail: undefined }
+  }
+
+  const seconds = duration.asSeconds()
+  if (seconds < 60) {
+    const decimals = seconds >= 10 ? 1 : 2
+    return {
+      display: `${sign}${seconds.toFixed(decimals)} s`,
+      detail: `${numberFormatter.format(value)} ms`,
+    }
+  }
+
+  const minutes = duration.asMinutes()
+  if (minutes < 60) {
+    const roundedSeconds = Math.round(seconds)
+    return {
+      display: `${sign}${minutes.toFixed(minutes >= 10 ? 1 : 2)} min`,
+      detail: `${numberFormatter.format(roundedSeconds)} s`,
+    }
+  }
+
+  const hours = duration.asHours()
+  const roundedMinutes = Math.round(minutes)
+  return {
+    display: `${sign}${hours.toFixed(hours >= 10 ? 1 : 2)} h`,
+    detail: `${numberFormatter.format(roundedMinutes)} min`,
+  }
+}
+
+export const getFormattedLagValue = (type: 'bytes' | 'duration', value?: number) =>
+  type === 'bytes' ? formatLagBytesValue(value) : formatLagDurationValue(value)
