+++ apps/studio/components/interfaces/Reports/UpdateModal.tsx
@@ -44,7 +44,6 @@ export const UpdateCustomReportModal = ({
         id: selectedReport.id,
         name: newVals.name,
         description: newVals.description || '',
-        favorite: selectedReport.favorite === null ? undefined : selectedReport.favorite,
       },
     })
   }
